Feat/day 6

+++ src/main/kotlin/Day3.kt
@@ -2,7 +2,7 @@ import Reader.readInputs
 
 class Day3A : Day() {
 
-    override fun run() : Int{
+    override fun run(): Long {
         val bits = retrieveBits()
         var gammaBits = ""
 
@@ -22,7 +22,7 @@ class Day3A : Day() {
         val epsilon = toEpsilon(gammaBits)
         val gamma = Integer.parseInt(gammaBits, 2)
 
-        return gamma * epsilon
+        return (gamma * epsilon).toLong()
     }
 
     private fun toEpsilon(gammaBits: String): Int {
@@ -43,7 +43,7 @@ class Day3A : Day() {
 
 class Day3B : Day() {
 
-    override fun run():Int {
+    override fun run(): Long {
         val bits = retrieveBits()
         val oxygenGeneratorRating = searchNumber(0, bits, true).let {
             Integer.parseInt(it, 2)
@@ -51,7 +51,7 @@ class Day3B : Day() {
         val c02ScrubberRating = searchNumber(0, bits, false).let {
             Integer.parseInt(it, 2)
         }
-        return oxygenGeneratorRating * c02ScrubberRating
+        return (oxygenGeneratorRating * c02ScrubberRating).toLong()
     }
 
     private fun searchNumber(currentIndex: Int, bits: List<String>, isMostCommonBitCriteria: Boolean): String {
